feat: Updated PE scope to create the PE next to the primary resource by default (#4449)
## Description

Background:
https://github.com/Azure/bicep-registry-modules/issues/3835#issuecomment-2658956734
Linked to: https://github.com/Azure/Azure-Verified-Modules/pull/1857

Changes the deployment so that the main resource's (e.g., Key Vaults) RG
is used as the default location for the PE. The already implemented
`resourceGroupResourceId` will continue to allow the user to specify a
different RG (in a different subscription, if needed).

The primary change is from
```bicep
    scope: resourceGroup(
      split(privateEndpoint.?resourceGroupResourceId ?? privateEndpoint.?subnetResourceId, '/')[2],
      split(privateEndpoint.?resourceGroupResourceId ?? privateEndpoint.?subnetResourceId, '/')[4]
    )
```
to
```bicep
    scope: resourceGroup(
      split(privateEndpoint.?resourceGroupResourceId ?? resourceGroup().id, '/')[2],
      split(privateEndpoint.?resourceGroupResourceId ?? resourceGroup().id, '/')[4]
    )
```

I'll quote from @ahmadabdalla on this matter
> In most scenarios, PEs are deployed alongside their main resource in
their own RG vs. the VNET RG. Customer app teams may have subnet join
permissions on a centralised VNET in a Landing zone, but may not have
permissions to deploy into it. Also considering billing and resource
lifecycle perspective.

cc: @JamesDawson

Ref: #4449

## Pipeline Reference

<!-- Insert your Pipeline Status Badge below -->

| Pipeline |
| -------- |
|          |

## Type of Change

<!-- Use the checkboxes [x] on the options that are relevant. -->

- [ ] Update to CI Environment or utilities (Non-module affecting
changes)
- [ ] Azure Verified Module updates:
- [ ] Bugfix containing backwards-compatible bug fixes, and I have NOT
bumped the MAJOR or MINOR version in `version.json`:
- [ ] Someone has opened a bug report issue, and I have included "Closes
#{bug_report_issue_number}" in the PR description.
- [ ] The bug was found by the module author, and no one has opened an
issue to report it yet.
- [ ] Feature update backwards compatible feature updates, and I have
bumped the MINOR version in `version.json`.
- [x] Breaking changes and I have bumped the MAJOR version in
`version.json`.
  - [ ] Update to documentation